update to windows compatability

+++ src/libs/libutil/utility/Utility.hpp
@@ -70,7 +70,7 @@ QString humanReadableElapsedMS(qint64 ms,int prec=2, bool prep=true,long double 
 QString formattedDateFromSeconds(long double seconds);
 QString formattedDateFromMS(qint64 ms);
 QString humanReadableSize(qint64 bytes,int prec=2);
-int snprint_hex(char *dst, size_t size, const uint8_t *pbtData, const size_t szBytes);
+size_t snprint_hex(char *dst, size_t size, const uint8_t *pbtData, const size_t szBytes);
 QString toHEX(const uint8_t *pbtData, const size_t szBytes);
 QString hexdump(void *mem, unsigned int len, QString topic="");
 QString humanReadablize(QString orig, int space=4, QString sep="-");
@@ -91,14 +91,14 @@ QString currentThreadID();
 class HandleCounter
 {
 private:
-	quint64 handleCounter;
-	QMap<Qt::HANDLE, quint64> handleMap;
-	QMutex handleMutex;
+    quint64 handleCounter;
+    QMap<Qt::HANDLE, quint64> handleMap;
+    QMutex handleMutex;
 
 public:
 
-	HandleCounter();
-	QString handleCounterString(Qt::HANDLE h);
+    HandleCounter();
+    QString handleCounterString(Qt::HANDLE h);
 
 };
 
@@ -164,33 +164,33 @@ void merge( QVariantMap  &c, QVariantMap  &b);
 // FROM https://stackoverflow.com/questions/21646467/how-to-execute-a-functor-or-a-lambda-in-a-given-thread-in-qt-gcd-style/21653558#21653558
 template <typename F>
 struct FunctorEvent : public QEvent {
-	using Functor = typename std::decay<F>::type;
-	Functor functor;
-	explicit FunctorEvent(Functor && fun) : QEvent(QEvent::None), functor(std::move(fun)) {}
-	explicit FunctorEvent(const Functor & fun) : QEvent(QEvent::None), functor(fun) {}
-	virtual ~FunctorEvent()
-	{
-		qDebug()<<"CALLING FUNCTOR FOR FUNCTOR EVENT";
-		functor();
-		qDebug()<<"DONE CALLING FUNCTOR FOR FUNCTOR EVENT";
-	}
+    using Functor = typename std::decay<F>::type;
+    Functor functor;
+    explicit FunctorEvent(Functor && fun) : QEvent(QEvent::None), functor(std::move(fun)) {}
+    explicit FunctorEvent(const Functor & fun) : QEvent(QEvent::None), functor(fun) {}
+    virtual ~FunctorEvent()
+    {
+        qDebug()<<"CALLING FUNCTOR FOR FUNCTOR EVENT";
+        functor();
+        qDebug()<<"DONE CALLING FUNCTOR FOR FUNCTOR EVENT";
+    }
 };
 
 template <typename F>
 static void postToObject(F && functor, QObject * obj = qApp)
 {
-	if (qobject_cast<QThread*>(obj)) {
-		qWarning() << "posting a call to a thread object - consider using postToThread";
-	}
-	QCoreApplication::postEvent(obj, OC_NEW FunctorEvent<F>(std::forward<F>(functor)));
+    if (qobject_cast<QThread*>(obj)) {
+        qWarning() << "posting a call to a thread object - consider using postToThread";
+    }
+    QCoreApplication::postEvent(obj, OC_NEW FunctorEvent<F>(std::forward<F>(functor)));
 }
 
 template <typename F>
 static void postToThread(F && functor, QThread * thread = qApp->thread())
 {
-	QObject * obj = QAbstractEventDispatcher::instance(thread);
-	Q_ASSERT(obj);
-	QCoreApplication::postEvent(obj, OC_NEW FunctorEvent<F>(std::forward<F>(functor)));
+    QObject * obj = QAbstractEventDispatcher::instance(thread);
+    Q_ASSERT(obj);
+    QCoreApplication::postEvent(obj, OC_NEW FunctorEvent<F>(std::forward<F>(functor)));
 }
 
 
@@ -199,50 +199,50 @@ static void postToThread(F && functor, QThread * thread = qApp->thread())
 class Q_CORE_EXPORT MutexTryLocker
 {
 private:
-	Q_DISABLE_COPY(MutexTryLocker)
-	quintptr val;
+    Q_DISABLE_COPY(MutexTryLocker)
+    quintptr val;
 public:
 
-	inline explicit MutexTryLocker(QBasicMutex *m) QT_MUTEX_LOCK_NOEXCEPT {
-		Q_ASSERT_X((reinterpret_cast<quintptr>(m) & quintptr(1u)) == quintptr(0), "MutexTryLocker", "QMutex pointer is misaligned");
-		val = quintptr(m);
-		if (Q_LIKELY(m))
-		{
-			// call QMutex::lock() instead of QBasicMutex::lock()
-			const bool is_locked=static_cast<QMutex *>(m)->try_lock();
-			if(is_locked) {
-				val |= 1;
-			}
-		}
-	}
-	inline ~MutexTryLocker()
-	{
-		unlock();
-	}
+    inline explicit MutexTryLocker(QBasicMutex *m) QT_MUTEX_LOCK_NOEXCEPT {
+        Q_ASSERT_X((reinterpret_cast<quintptr>(m) & quintptr(1u)) == quintptr(0), "MutexTryLocker", "QMutex pointer is misaligned");
+        val = quintptr(m);
+        if (Q_LIKELY(m))
+        {
+            // call QMutex::lock() instead of QBasicMutex::lock()
+            const bool is_locked=static_cast<QMutex *>(m)->try_lock();
+            if(is_locked) {
+                val |= 1;
+            }
+        }
+    }
+    inline ~MutexTryLocker()
+    {
+        unlock();
+    }
 
-	inline void unlock() Q_DECL_NOTHROW {
-		if ((val & quintptr(1u)) == quintptr(1u))
-		{
-			val &= ~quintptr(1u);
-			mutex()->unlock();
-		}
-	}
+    inline void unlock() Q_DECL_NOTHROW {
+        if ((val & quintptr(1u)) == quintptr(1u))
+        {
+            val &= ~quintptr(1u);
+            mutex()->unlock();
+        }
+    }
 
-	inline void relock() QT_MUTEX_LOCK_NOEXCEPT {
-		if (val)
-		{
-			if ((val & quintptr(1u)) == quintptr(0u)) {
-				mutex()->lock();
-				val |= quintptr(1u);
-			}
-		}
-	}
+    inline void relock() QT_MUTEX_LOCK_NOEXCEPT {
+        if (val)
+        {
+            if ((val & quintptr(1u)) == quintptr(0u)) {
+                mutex()->lock();
+                val |= quintptr(1u);
+            }
+        }
+    }
 
 
-	inline QMutex *mutex() const
-	{
-		return reinterpret_cast<QMutex *>(val & ~quintptr(1u));
-	}
+    inline QMutex *mutex() const
+    {
+        return reinterpret_cast<QMutex *>(val & ~quintptr(1u));
+    }
 
 
 
@@ -255,18 +255,18 @@ public:
 template <typename T>
 void pack(T *stacked)
 {
-	stacked->updatesEnabled();
-	stacked->setUpdatesEnabled(false);
-	for(int i=0; i<stacked->count(); i++) {
-		stacked->widget(i)->setSizePolicy(QSizePolicy::Ignored, QSizePolicy::Ignored);
-	}
-	QWidget *widget=stacked->currentWidget();
-	widget->setSizePolicy(QSizePolicy::Minimum, QSizePolicy::Minimum);
-	widget->resize(widget->minimumSizeHint());
-	widget->adjustSize();
-	stacked->resize(stacked->minimumSizeHint());
-	stacked->adjustSize();
-	stacked->setUpdatesEnabled(true);
+    stacked->updatesEnabled();
+    stacked->setUpdatesEnabled(false);
+    for(int i=0; i<stacked->count(); i++) {
+        stacked->widget(i)->setSizePolicy(QSizePolicy::Ignored, QSizePolicy::Ignored);
+    }
+    QWidget *widget=stacked->currentWidget();
+    widget->setSizePolicy(QSizePolicy::Minimum, QSizePolicy::Minimum);
+    widget->resize(widget->minimumSizeHint());
+    widget->adjustSize();
+    stacked->resize(stacked->minimumSizeHint());
+    stacked->adjustSize();
+    stacked->setUpdatesEnabled(true);
 }
